# gh-page/src/components/swagger/api.yaml
openapi: 3.0.0

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

security:
  - BearerAuth: []

info:
  title: Glasswall File Rebuild Product API
  description: These are HTTP endpoints that rebuild files with the Glasswall d-FIRST™ Engine.
  version: 0.1.9
servers:
  - url: https://gzlhbtpvk2.execute-api.eu-west-1.amazonaws.com/Prod
    description: eu-west-1
paths:
  /api/rebuild:
    post:
      summary: Rebuilds a file located at a specified URL and outputs the rebuilt file to another specified URL
      description: This endpoint accepts requests to rebuild a file with Glasswall d-FIRST™ Engine. The request body contains the Base64 representation of the file and Glasswall Content Management Flags with a 'Content-Type' of 'application/json'. A Base64 Representation of the rebuilt file is then returned in the response with a 'Content-Type' of 'text/plain'.
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              required:
                - InputGetUrl
                - OutputPutUrl
              properties:
                InputGetUrl:
                  type: string
                  description: This URL will be used in a HTTP GET operation to download the file to be rebuilt
                OutputPutUrl:
                  type: string
                  description: This URL will be used upon successful rebuilding of a file in a HTTP PUT operation. The ETAG of the PUT file will be returned in the headers (See responses)
                ContentManagementFlags:
                  $ref: '#/Models/ContentManagementPolicy'
            encoding:
              file:
                style: form
      responses:
        200:    # status code
          description: OK - The requested file was successfully processed by the Glasswall d-FIRST™ Engine
          headers:
            gw-metric-detect:
              description: A header containing the time it took to detect the file type.
              schema:
                type: string
            gw-metric-filesize:
              description: A header containing a record of the size of the file uploaded
              schema:
                type: string
            gw-version:
              description: A header containing the version of the Glasswall d-FIRST™ Engine installed on the server
              schema:
                type: string
            gw-metric-rebuild:
              description: A header containing the time it took to rebuild the input file.
              schema:
                type: string
            gw-metric-download:
              description: A header containing the time it took to download the input file
              schema:
                type: string
            gw-metric-upload:
              description: A header containing the time it took to upload the rebuilt file
              schema:
                type: string
            gw-metric-uploadsize:
              description: A header containing the size of the rebuilt file
              schema:
                type: string
            gw-put-file-etag:
              description: A header containing the etag of the output file. 
              schema:
                type: string
        400:
          description: Bad Request - This is usually when the JSON input is malformed or missing parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  InputGetUrl:
                    type: array
                    description: "Error relating to the input"
                    items:
                      type: string
                  OutputPutUrl:
                    type: array
                    description: "Error relating to the input"
                    items:
                      type: string
        403:
          $ref: '#/responses/Forbidden' 
        415:
          $ref: '#/responses/UnsupportedMediaType' 
        422:
          $ref: '#/responses/UnprocessableEntity' 
  /api/rebuild/file:
    post:
      summary: Rebuilds a file using its binary data
      description: <h3>Description:</h3>This endpoint accepts requests to rebuild a file with Glasswall d-FIRST™ Engine.  Both the file and the Content Management Policy are sent in the request body with a 'Content-Type' of 'multipart/form-data'. The Rebuilt file is then returned in the response body with a 'Content-Type' of 'application/octet-stream'.
      parameters:
        - in: header
          name: Accept
          description: A value of 'application/octet-stream' will return the rebuilt file's binary content to the client. If this is not set then the file's Base64 representation will be returned instead.
          schema:
            type: string
            default: application/octet-stream
          required: false
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: The binary contents of the selected file will be uploaded to the server. The Total supported request size of the API gateway is 6MB, therefore the file contents must also be less than 6MB.
                contentManagementFlagJson:
                  $ref: '#/Models/ContentManagementPolicy'
              required:
                - file
            encoding:
              file:
                style: form
              contentManagementFlagJson:
                contentType: application/json
                style: form
      responses:
        200:    # status code
          description: OK - The requested file was processed by the Glasswall d-FIRST™ Engine
          content: 
            application/octet-stream:
              schema:
                format: binary
                description: The body is a binary stream, containing the rebuilt file
            application/json:
              schema:
                description: This is returned when the request was successful but the content management policy has is set to disallow content found in the file.
                type: object
                properties:
                  IsDisallowed:
                    type: boolean
                    description: A value indicating whether the file was disallowed by the Content Management Policy
                  ErrorMessage:
                    type: string
                    description: A description of the reason the file was not protected.
          headers:
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename=ExampleFileName.pdf; filename*=UTF-8''ExampleFileName.pdf
              description: This contains the details of the rebuilt files binary data to download
            gw-metric-detect:
              description: A header containing the time it took to detect the file type.
              schema:
                type: string
            gw-metric-filesize:
              description: A header containing a record of the size of the file uploaded
              schema:
                type: string
            gw-version:
              description: A header containing the version of the Glasswall d-FIRST™ Engine installed on the server
              schema:
                type: string
            gw-metric-rebuild:
              description: A header containing the time it took to rebuild the input file.
              schema:
                type: string
            gw-metric-formfileread:
              description: A header containing the time it took to read the file uploaded
              schema:
                type: string
        400:
          description: Bad Request - This is usually when the JSON input is malformed or missing parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  file:
                    type: array
                    description: "Error relating to the input file"
                    items:
                      type: string
        403:
          $ref: '#/responses/Forbidden' 
        415:
          $ref: '#/responses/UnsupportedMediaType' 
        422:
          $ref: '#/responses/UnprocessableEntity' 
          description: This occurs when the request was understood, but the file requested could not be rebuilt by the engine. The response body contains the reason.
  /api/rebuild/base64:
    post:
      summary: Rebuilds a file using the Base64 encoded representation
      description: >
        This endpoint accepts requests to rebuild a file with Glasswall d-FIRST™ Engine.
        The request body contains the Base64 representation of the file and Glasswall Content Management Flags with a 'Content-Type' of 'application/json'. 
        A Base64 Representation of the rebuilt file is then returned in the response with a 'Content-Type' of 'text/plain'.
        <br/><br/>
        Select a file below to copy its Base64 Encoded representation to clipboard. The Total supported request size of the API gateway is 6MB, therefore the base64 encoded string must also be less than 6MB.
        <br/><br/>
        <span><input id="base64Input" type="file"></input> 
      requestBody:
        content:
          "application/json":
            schema:
              type: object
              required:
                - Base64
              properties:
                Base64:
                  type: string
                  description: This is the Base64 encoded representation of the file to rebuild. 
                ContentManagementFlags:
                  $ref: '#/Models/ContentManagementPolicy'
            encoding:
              Base64:
                style: form
      responses:
        200:    # status code
          description: OK - The requested Base64 was successfully decoded and processed by the Glasswall d-FIRST™ Engine
          content: 
            text/plain:
              schema:
                description: The body is a rebuilt file, Base64 encoded.
                type: string
                format: base64
                example: "0M8R4KGxGuEAAAAAAAAAAAAAAAAAAAAAPgADAP7/CQAGAAAAAAAAAAAAAAABAAAAUgAAAAAAAAAAEAAAAAAAAAEAAAD+////AAAAAFMAAAD//////////////////////////////..."
            application/json:
              schema:
                description: This is returned when the request was successful but the content management policy has is set to disallow content found in the file.
                type: object
                properties:
                  IsDisallowed:
                    type: boolean
                    description: A value indicating whether the file was disallowed by the Content Management Policy
                  ErrorMessage:
                    type: string
                    description: A description of the reason the file was not protected.
          headers:
            gw-metric-detect:
              description: A header containing the time it took to detect the file type.
              schema:
                type: string
            gw-metric-filesize:
              description: A header containing a record of the size of the file uploaded
              schema:
                type: string
            gw-version:
              description: A header containing the version of the Glasswall d-FIRST™ Engine installed on the server
              schema:
                type: string
            gw-metric-rebuild:
              description: A header containing the time it took to rebuild the input file.
              schema:
                type: string
            gw-metric-decode-base64:
              description: A header containing the time it took to read decode the Base64 string
              schema:
                type: string
        400:
          description: Bad Request - This is usually when the JSON input is malformed or missing parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  Base64:
                    type: array
                    description: "Error relating to the input"
                    items:
                      type: string
        403:
          $ref: '#/responses/Forbidden' 
        415:
          $ref: '#/responses/UnsupportedMediaType' 
        422:
          $ref: '#/responses/UnprocessableEntity' 
responses:
  Forbidden:
    description: Forbidden - This typically occurs when the JWT Token is not supplied, or it is incorrect.
    content:
      "application/json":
        schema:
          type: object
          properties:
            message:
              type: string
  UnsupportedMediaType:
    description: Unsupported Media Type - This happens when the request was not sent in JSON.
  UnprocessableEntity:
    description: Unprocessable Entity - This occurs when the Glasswall Engine was unable to rebuild the file due to an unsupported file type or if it is non conforming.
Models:
  ContentManagementPolicy:
    type: object
    description: >
      This field contains each of the Content Management Flags for the file types that the engine supports. The server treats this field as a JSON string. All the properties including the field itself are optional. 
              <br/><br/> Content Management Flag Key:
              <br/><br/> `0` - Allow
              <br/><br/> `1` - Sanitise (Default)
              <br/><br/> `2` - Disallow
    properties: 
      PdfContentManagement:
        type: object
        properties:
          Metadata:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for PDF Metadata, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          InternalHyperlinks:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for PDF Interal Hyperlinks, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          ExternalHyperlinks:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for PDF External Hyperlinks, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          EmbeddedFiles:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for files Embedded in PDF, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          EmbeddedImages:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for PDF Embedded Images, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          Javascript:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for PDF Javascript code, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          Acroform:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for PDF Acroforms, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          ActionsAll:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for PDF Actions, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
      ExcelContentManagement:
        type: object
        properties:
          Metadata:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Excel Metadata, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          InternalHyperlinks:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Excel Internal Hyperlinks, valid values:
              * `0` - Allow
              * `1` - Sanitise (Default)
              * `2` - Disallow
          ExternalHyperlinks:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Excel External Hyperlinks, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          EmbeddedFiles:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for files Embedded in Excel documents, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          EmbeddedImages:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for images Embedded in Excel documents, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          DynamicDataExchange:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Excel Dynamic Data, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          Macros:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Excel Macros, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          ReviewComments:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Excel Review Comments, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
      PowerPointContentManagement:
        type: object
        properties:
          Metadata:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Powerpoint Metadata, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          InternalHyperlinks:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Powerpoint Internal Hyperlinks, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          ExternalHyperlinks:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Powerpoint External Hyperlinks, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          EmbeddedFiles:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Powerpoint Embedded Files, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          EmbeddedImages:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Powerpoint Embedded Images, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          Macros:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Powerpoint Macros, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          ReviewComments:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Powerpoint Review Comments, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
      WordContentManagement:
        type: object
        properties:
          Metadata:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Word Documents Metadata, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          InternalHyperlinks:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Word Documents Internal Hyperlinks, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          ExternalHyperlinks:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Word Documents External Hyperlinks, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          EmbeddedFiles:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Word Documents Embedded Files, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          EmbeddedImages:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Word Documents Embedded Images, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          DynamicDataExchange:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Word Documents Dynamic Data, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          Macros:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Word Documents Macros, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow
          ReviewComments:
            type: integer
            enum: [1, 0, 2]
            description: >
              Specifies the content action for Word Documents Review Comments, valid values:
              <br/><br/>* `0` - Allow
              <br/><br/>* `1` - Sanitise (Default)
              <br/><br/>* `2` - Disallow